requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.11
mysqlclient==2.2.7
numpy==2.4.2
opencv-python==4.13.0.90
pillow==12.1.0
PyJWT==2.11.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
